pyproject.toml
[project]
name = "a2a-mcp-tutorial"
version = "1.0.0"
description = "A tutorial on how to use Model Context Protocol by Anthropic and Agent2Agent Protocol by Google"
requires-python = ">=3.12"
dependencies = [
    "finnhub-python>=2.4.23",
    "google-adk>=0.1.0",
    "httpx>=0.28.1",
    "litellm>=1.30.0",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "termcolor>=3.0.1",
    "yfinance>=0.2.61",
]
